moved useFindOne to own file
